refactor(backend): sink DataTableController logic into services
Thin out DataTableController (1187 -> ~520 lines), moving business
orchestration into the service layer:
- DataTableService: requireTableLocation (single table-identity resolver,
  replacing 7 duplicated dbName/tableName blocks) + lifecycle write
  commands (updateTable / updateTableComment / softDeleteTable /
  restoreTable / purgeTableNow / listPendingDeletionView) + id-based
  field overloads; @Transactional facades preserve inner-write atomicity
  under self-invocation.
- DataTableQueryService (new): statistics(+cache/history) / access-stats /
  ddl / preview / export / database-statistics / history read orchestration.
- DataTableMetadataSyncService (new): audit + sync(all/db/by-name/by-id),
  history recording, response assembly, tableId enrichment; preserves the
  'sync exception captured as FAIL result, still returned' semantics.
- New value types TableLocation, TableExport.

Controller now only binds requests, keeps @RequireAuth, shapes HTTP/Result
responses and maps exceptions via thin try/catch (deps 9 -> 3). Duplicated
private helpers removed; delete() kept thin (shared static confirm helper).

Behavior-preserving except two documented consolidations (see design
Risks): a few previously-uncaught failures now return 200/code500/raw like
the dominant path; lifecycle Doris DDL runs within the facade transaction.

Tests: DataTableControllerTest updated to new constructor + delegation;
DataTableMetadataSyncServiceTest covers history/ tableId / clusterId guard.
