+++ backend/src/test/java/com/onedata/portal/controller/DataTableControllerTest.java
@@ -1,17 +1,8 @@
 package com.onedata.portal.controller;
 
-import com.onedata.portal.entity.DataTable;
-import com.onedata.portal.entity.DorisCluster;
-import com.onedata.portal.entity.MetadataSyncHistory;
-import com.onedata.portal.service.DataExportService;
+import com.onedata.portal.service.DataTableMetadataSyncService;
+import com.onedata.portal.service.DataTableQueryService;
 import com.onedata.portal.service.DataTableService;
-import com.onedata.portal.service.DorisClusterService;
-import com.onedata.portal.service.DorisConnectionService;
-import com.onedata.portal.service.DorisMetadataSyncService;
-import com.onedata.portal.service.DorisTableAccessService;
-import com.onedata.portal.service.MetadataSyncHistoryService;
-import com.onedata.portal.service.TableStatisticsCacheService;
-import com.onedata.portal.service.TableStatisticsHistoryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -20,12 +11,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import java.time.LocalDateTime;
+import java.util.HashMap;
+import java.util.Map;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -38,28 +29,10 @@ class DataTableControllerTest {
     private DataTableService dataTableService;
 
     @Mock
-    private DorisConnectionService dorisConnectionService;
+    private DataTableQueryService dataTableQueryService;
 
     @Mock
-    private TableStatisticsCacheService cacheService;
-
-    @Mock
-    private TableStatisticsHistoryService historyService;
-
-    @Mock
-    private DataExportService dataExportService;
-
-    @Mock
-    private DorisMetadataSyncService dorisMetadataSyncService;
-
-    @Mock
-    private DorisTableAccessService dorisTableAccessService;
-
-    @Mock
-    private DorisClusterService dorisClusterService;
-
-    @Mock
-    private MetadataSyncHistoryService metadataSyncHistoryService;
+    private DataTableMetadataSyncService dataTableMetadataSyncService;
 
     private MockMvc mockMvc;
 
@@ -67,41 +40,22 @@ class DataTableControllerTest {
     void setUp() {
         DataTableController controller = new DataTableController(
                 dataTableService,
-                dorisConnectionService,
-                cacheService,
-                historyService,
-                dataExportService,
-                dorisMetadataSyncService,
-                dorisTableAccessService,
-                dorisClusterService,
-                metadataSyncHistoryService);
+                dataTableQueryService,
+                dataTableMetadataSyncService);
         mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
     }
 
     @Test
-    void syncTableMetadataByNameReturnsSyncedTableId() throws Exception {
-        DorisCluster cluster = new DorisCluster();
-        cluster.setId(1L);
-        cluster.setClusterName("local-doris");
-        cluster.setSourceType("DORIS");
+    void syncTableMetadataByNameDelegatesAndMapsSuccessMessage() throws Exception {
+        Map<String, Object> response = new HashMap<>();
+        response.put("success", true);
+        response.put("status", "SUCCESS");
+        response.put("syncRunId", 77L);
+        response.put("database", "dw");
+        response.put("tableName", "fact_orders");
+        response.put("tableId", 42L);
 
-        DorisMetadataSyncService.SyncResult syncResult = new DorisMetadataSyncService.SyncResult();
-        syncResult.addNewTable();
-
-        MetadataSyncHistory history = new MetadataSyncHistory();
-        history.setId(77L);
-
-        DataTable syncedTable = new DataTable();
-        syncedTable.setId(42L);
-        syncedTable.setClusterId(1L);
-        syncedTable.setDbName("dw");
-        syncedTable.setTableName("fact_orders");
-
-        when(dorisClusterService.getById(1L)).thenReturn(cluster);
-        when(dorisMetadataSyncService.syncTable(1L, "dw", "fact_orders")).thenReturn(syncResult);
-        when(metadataSyncHistoryService.record(eq(cluster), eq("manual"), eq("table"), eq("dw.fact_orders"),
-                any(LocalDateTime.class), eq(syncResult))).thenReturn(history);
-        when(dataTableService.getByDbAndTableName(1L, "dw", "fact_orders")).thenReturn(syncedTable);
+        when(dataTableMetadataSyncService.syncTableByName(1L, "dw", "fact_orders")).thenReturn(response);
 
         mockMvc.perform(post("/v1/tables/sync-metadata/database/dw/table/fact_orders")
                         .param("clusterId", "1"))
@@ -114,18 +68,17 @@ class DataTableControllerTest {
                 .andExpect(jsonPath("$.data.tableName").value("fact_orders"))
                 .andExpect(jsonPath("$.data.tableId").value(42));
 
-        verify(dorisMetadataSyncService).syncTable(1L, "dw", "fact_orders");
-        verify(metadataSyncHistoryService).record(eq(cluster), eq("manual"), eq("table"), eq("dw.fact_orders"),
-                any(LocalDateTime.class), eq(syncResult));
+        verify(dataTableMetadataSyncService).syncTableByName(1L, "dw", "fact_orders");
     }
 
     @Test
-    void syncTableMetadataByNameRequiresClusterId() throws Exception {
+    void syncTableMetadataByNameSurfacesMissingClusterIdMessage() throws Exception {
+        when(dataTableMetadataSyncService.syncTableByName(isNull(), eq("dw"), eq("fact_orders")))
+                .thenThrow(new RuntimeException("请指定数据源"));
+
         mockMvc.perform(post("/v1/tables/sync-metadata/database/dw/table/fact_orders"))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.code").value(500))
                 .andExpect(jsonPath("$.message").value("请指定数据源"));
-
-        verifyNoInteractions(dorisMetadataSyncService);
     }
 }
